# [WIP] Lessen load of GC