--- a/lib/content/local/local.go
+++ b/lib/content/local/local.go
@@ -10,16 +10,16 @@ import (
 	"github.com/livesense-inc/fanlin/lib/content"
 )
 
-func GetImageBinary(c *content.Content, buf *bytes.Buffer) (io.Reader, error) {
+func GetImageBinary(c *content.Content, b *bytes.Buffer) (io.Reader, error) {
 	f, err := os.Open(path.Clean(c.SourcePlace))
 	if err != nil {
 		return nil, fmt.Errorf("failed to open a file: %s: %w", c.SourcePlace, err)
 	}
 	defer f.Close()
-	if _, err := io.Copy(buf, f); err != nil {
+	if _, err := io.Copy(b, f); err != nil {
 		return nil, err
 	}
-	return buf, nil
+	return b, nil
 }
 
 func init() {