--- a/lib/content/s3/s3.go
+++ b/lib/content/s3/s3.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
-	s3manager "github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/livesense-inc/fanlin/lib/content"
 	imgproxyerr "github.com/livesense-inc/fanlin/lib/error"
@@ -19,11 +18,11 @@ import (
 var s3GetSourceFunc = getS3ImageBinary
 
 // Test dedicated function
-func setS3GetFunc(f func(cfg *aws.Config, bucket, key string, b []byte) (io.Reader, error)) {
+func setS3GetFunc(f func(cfg *aws.Config, bucket, key string, b *bytes.Buffer) (io.Reader, error)) {
 	s3GetSourceFunc = f
 }
 
-func GetImageBinary(c *content.Content, b []byte) (io.Reader, error) {
+func GetImageBinary(c *content.Content, b *bytes.Buffer) (io.Reader, error) {
 	if c == nil {
 		return nil, errors.New("content is nil")
 	}
@@ -69,18 +68,21 @@ func NormalizePath(path string, form string) (string, error) {
 	return "", imgproxyerr.New(imgproxyerr.WARNING, errors.New("invalid normalization form("+form+")"))
 }
 
-func getS3ImageBinary(cfg *aws.Config, bucket, key string, b []byte) (io.Reader, error) {
-	downloader := s3manager.NewDownloader(s3.NewFromConfig(*cfg))
-	buf := s3manager.NewWriteAtBuffer(b)
-	input := &s3.GetObjectInput{
+func getS3ImageBinary(cfg *aws.Config, bucket, key string, buf *bytes.Buffer) (io.Reader, error) {
+	cli := s3.NewFromConfig(*cfg)
+	input := s3.GetObjectInput{
 		Bucket: aws.String(bucket),
 		Key:    aws.String(key),
 	}
-	_, err := downloader.Download(context.TODO(), buf, input)
+	output, err := cli.GetObject(context.TODO(), &input)
 	if err != nil {
 		return nil, imgproxyerr.New(imgproxyerr.WARNING, err)
 	}
-	return bytes.NewReader(buf.Bytes()), imgproxyerr.New(imgproxyerr.ERROR, err)
+	defer output.Body.Close()
+	if _, err := io.Copy(buf, output.Body); err != nil {
+		return nil, imgproxyerr.New(imgproxyerr.WARNING, err)
+	}
+	return buf, nil
 }
 
 func init() {