--- a/lib/content/s3/s3_test.go
+++ b/lib/content/s3/s3_test.go
@@ -24,7 +24,7 @@ func initialize() {
 	testKey = "test/test.jpg"
 }
 
-func mockS3GetFunc(config *aws.Config, bucket, key string) (io.Reader, error) {
+func mockS3GetFunc(config *aws.Config, bucket, key string, b []byte) (io.Reader, error) {
 	if config == nil {
 		return strings.NewReader("failed"), errors.New("config is empty")
 	} else if config.Region != testRegion {
@@ -54,11 +54,11 @@ func newTestContent() *content.Content {
 func TestGetImageBinary(t *testing.T) {
 	initialize()
 	c := newTestContent()
-	if _, err := GetImageBinary(c); err != nil {
+	if _, err := GetImageBinary(c, []byte{}); err != nil {
 		t.Log("normal pattern.")
 		t.Fatal(err)
 	}
-	if _, err := GetImageBinary(nil); err == nil {
+	if _, err := GetImageBinary(nil, []byte{}); err == nil {
 		t.Log("abnormal pattern.")
 		t.Fatal("err is nil.")
 	}