--- a/lib/content/source.go
+++ b/lib/content/source.go
@@ -10,14 +10,14 @@ import (
 
 type source struct {
 	name           string
-	getImageBinary func(*Content) (io.Reader, error)
+	getImageBinary func(*Content, []byte) (io.Reader, error)
 }
 
 var sources []source
 
 // RegisterContentType registers an content type for use by GetContent.
 // Name is the name of the content type, like "web" or "s3".
-func RegisterContentType(name string, getImageBinary func(*Content) (io.Reader, error)) {
+func RegisterContentType(name string, getImageBinary func(*Content, []byte) (io.Reader, error)) {
 	sources = append(sources, source{
 		name,
 		getImageBinary,
@@ -34,12 +34,12 @@ func sniff(c *Content) source {
 	return source{}
 }
 
-func GetImageBinary(c *Content) (io.Reader, error) {
+func GetImageBinary(c *Content, b []byte) (io.Reader, error) {
 	f := sniff(c)
 	if f.getImageBinary == nil {
 		return nil, imgproxyerr.New(imgproxyerr.WARNING, errors.New("unknown content type"))
 	}
-	m, err := f.getImageBinary(c)
+	m, err := f.getImageBinary(c, b)
 	if err != nil {
 		return nil, err
 	}