--- a/lib/content/web/web.go
+++ b/lib/content/web/web.go
@@ -9,7 +9,7 @@ import (
 	"io"
 
 	"github.com/livesense-inc/fanlin/lib/content"
-	"github.com/livesense-inc/fanlin/lib/error"
+	imgproxyerr "github.com/livesense-inc/fanlin/lib/error"
 )
 
 var ua = fmt.Sprintf("Mozilla/5.0 (fanlin; arch: %s; OS: %s; Go version: %s) Go language Client/1.1 (KHTML, like Gecko) Version/1.0 fanlin", runtime.GOARCH, runtime.GOOS, runtime.Version())
@@ -22,14 +22,14 @@ type RealWebClient struct {
 }
 
 type WebClient interface {
-	Get(string) (io.Reader, error)
+	Get(string, []byte) (io.Reader, error)
 }
 
 type Client struct {
 	Http WebClient
 }
 
-func (r *RealWebClient) Get(url string) (io.Reader, error) {
+func (r *RealWebClient) Get(url string, b []byte) (io.Reader, error) {
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, imgproxyerr.New(imgproxyerr.ERROR, err)
@@ -46,7 +46,7 @@ func (r *RealWebClient) Get(url string) (io.Reader, error) {
 		return nil, imgproxyerr.New(imgproxyerr.WARNING, fmt.Errorf("received error status code(%d)", resp.StatusCode))
 	}
 
-	buffer := new(bytes.Buffer)
+	buffer := bytes.NewBuffer(b)
 	if _, err := io.Copy(buffer, resp.Body); err != nil {
 		return nil, err
 	}
@@ -63,8 +63,8 @@ func isErrorCode(status int) bool {
 	}
 }
 
-func GetImageBinary(c *content.Content) (io.Reader, error) {
-	return httpClient.Http.Get(c.SourcePlace)
+func GetImageBinary(c *content.Content, b []byte) (io.Reader, error) {
+	return httpClient.Http.Get(c.SourcePlace, b)
 }
 
 func setHttpClient(c Client) {