--- a/lib/image/image.go
+++ b/lib/image/image.go
@@ -190,7 +190,7 @@ func EncodeAVIF(buf io.Writer, img *image.Image, q int) error {
 }
 
 // DecodeImage is return image.Image
-func DecodeImage(r io.Reader, b []byte) (*Image, error) {
+func DecodeImage(r io.Reader, b *bytes.Buffer) (*Image, error) {
 	img, format, err := decode(r, b)
 	return &Image{img: img, format: format}, imgproxyerr.New(imgproxyerr.WARNING, err)
 }
@@ -311,7 +311,7 @@ func (i *Image) GetFormat() string {
 	return i.format
 }
 
-func Set404Image(buf io.Writer, tmpBuf []byte, path string, w uint, h uint, c color.Color, maxW uint, maxH uint) error {
+func Set404Image(buf io.Writer, tmpBuf *bytes.Buffer, path string, w uint, h uint, c color.Color, maxW uint, maxH uint) error {
 	f, err := os.Open(path)
 	if err != nil {
 		return imgproxyerr.New(imgproxyerr.ERROR, err)
@@ -363,8 +363,7 @@ func readOrientation(r io.Reader) (o int, err error) {
 	return
 }
 
-func decode(r io.Reader, b []byte) (d image.Image, format string, err error) {
-	buf := bytes.NewBuffer(b)
+func decode(r io.Reader, buf *bytes.Buffer) (d image.Image, format string, err error) {
 	tee := io.TeeReader(r, buf)
 
 	s, format, err := image.Decode(tee)