… lib/image/image_test.go