## Summery
* Get rid of all unwieldy handlings with panic to be able to use defer
* Use [sync.Pool](https://pkg.go.dev/sync#Pool) for [bytes.Buffer](https://pkg.go.dev/bytes#Buffer) to lessen the load of the GC
* Use [s3.Client](https://pkg.go.dev/github.com/aws/aws-sdk-go-v2/service/s3#Client.GetObject) instead of [s3manager.Downloader](https://pkg.go.dev/github.com/aws/aws-sdk-go-v2/feature/s3/manager#Downloader) to be able to use [bytes.Buffer](https://pkg.go.dev/bytes#Buffer)
  * There's still room for consideration. Downloader is concurrent.

## Benchmark
`B/op` was a bit decreased.

### Before
```
cpu: AMD EPYC 7763 64-Core Processor                
BenchmarkMainHandler-4   	read configure. : /home/runner/work/fanlin/fanlin/lib/test/test_conf9.json
      16	  66068110 ns/op	13052543 B/op	 1573239 allocs/op
```

### After
```
cpu: AMD EPYC 7763 64-Core Processor                
BenchmarkMainHandler-4   	read configure. : /home/runner/work/fanlin/fanlin/lib/test/test_conf9.json
      18	  64630002 ns/op	11019228 B/op	 1573223 allocs/op
```

## Profiling
`bytes.growSlice` and `bytes.(*Buffer).grow` were a bit decreased.

### Before
```
Showing top 10 nodes out of 45
      flat  flat%   sum%        cum   cum%
      74MB 33.93% 33.93%       74MB 33.93%  image.(*YCbCr).At
   68.50MB 31.41% 65.34%   142.50MB 65.34%  github.com/BurntSushi/graphics-go/graphics/interp.bilinearGeneral
   34.04MB 15.61% 80.95%    34.04MB 15.61%  bytes.growSlice
      24MB 11.00% 91.96%       24MB 11.00%  image.NewRGBA
   13.52MB  6.20% 98.16%    13.52MB  6.20%  image.NewYCbCr
    2.01MB  0.92% 99.08%     2.01MB  0.92%  io.ReadAll
         0     0% 99.08%    17.94MB  8.22%  bufio.(*Reader).Read
         0     0% 99.08%    16.11MB  7.39%  bytes.(*Buffer).ReadFrom
         0     0% 99.08%    17.94MB  8.22%  bytes.(*Buffer).Write
         0     0% 99.08%    34.04MB 15.61%  bytes.(*Buffer).grow
```

### After
```
Showing top 10 nodes out of 31
      flat  flat%   sum%        cum   cum%
   76.50MB 37.75% 37.75%   148.50MB 73.28%  github.com/BurntSushi/graphics-go/graphics/interp.bilinearGeneral
      72MB 35.53% 73.28%       72MB 35.53%  image.(*YCbCr).At
   32.66MB 16.11% 89.39%    32.66MB 16.11%  image.NewRGBA
   16.41MB  8.10% 97.49%    16.41MB  8.10%  image.NewYCbCr
    1.51MB  0.74% 98.23%     1.51MB  0.74%  github.com/rwcarlsen/goexif/exif.newAppSec
    1.07MB  0.53% 98.76%     1.07MB  0.53%  bytes.growSlice
    0.51MB  0.25% 99.01%    16.92MB  8.35%  image/jpeg.Decode
         0     0% 99.01%     1.07MB  0.53%  bytes.(*Buffer).ReadFrom
         0     0% 99.01%     1.07MB  0.53%  bytes.(*Buffer).grow
         0     0% 99.01%   148.50MB 73.28%  github.com/BurntSushi/graphics-go/graphics.Affine.Transform
```